rename & move MainActivity.java
to unread.UnreadItemsActivity.java
