+++ linen-starter/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:name="android.support.multidex.MultiDexApplication"
         android:theme="@style/AppTheme.Polyfill" >
         <activity
-            android:name=".MainActivity"
+            android:name=".unread.UnreadItemsActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
