+++ linen-starter/src/main/java/x7c1/linen/settings/MyChannelSourcesActivity.java
@@ -2,7 +2,7 @@ package x7c1.linen.settings;
 
 import android.os.Bundle;
 
-import x7c1.linen.BaseActivity;
+import x7c1.linen.base.BaseActivity;
 import x7c1.linen.R;
 import x7c1.linen.base.TransitAnimation;
 import x7c1.linen.base.TransitAnimations;
